Add id attributes to individual messages to permit more convenient references.
The id value is the message timestamp formatted in a style inspired by
the Clojure #inst literal. While basing the id on the timestamp could
potentially cause a conflict if two messages have the same timestamp value,
the code already assumes timestamp uniqueness so as written no conflict
can occur.

Change-Id: I6b11a882919672feeffe22d42b09cfcf6b47f518